add-pagination-auth-image-upload

+++ Backend/utils/fileDelete.js
@@ -1,9 +1,12 @@
 // this will help you delete files
 const fs = require("fs");
 const fileDelete = (path) => {
-  fs.unlink(path, (err) => {
-    if (err) throw new Error(err.message);
-  });
+  console.log(fs.existsSync(path));
+  if(fs.existsSync(path)) {
+    fs.unlink(path, (err) => {
+      if (err) throw new Error(err.message);
+    });
+  }
 };
 
 module.exports = fileDelete;
